tested finiteSizedFilms for postProcessing (spatial resolution), improved animation for higher resolution videos
Since the initial condition now returns a column vector and the previously written spatial resolution code considered it as a row vector, there was a transpose operation in that file that is no longer needed now
Animation file is now changed from avoiding a for loop (generated a lot of empty frames depending on the value of animationSkip) to a while loop.
Furthermore, it allows a handle on the resolution which can be changed from the current value of 150 to higher values.