disjoining pressure switch
added a switch to include van der Waals forces in the calculation.
Also changed the cut-off from 2e-5 to 0.05, since it takes a lot of time for the film to reach that thickness of 2e-5, and also 2e-3, and 0.01 (atleast took 1800 sec to finish the simulation and  9000 sec to make an animation). Hence changed it to 0.05.